adding code for errors check
