--- src/components/centerblock_components/filter/utils.js
@@ -1,11 +1,0 @@
-export const getFilter = (tracks, { id, filterBy, filterByKey }) => {
-  const filter = {
-    id,
-    filterBy,
-    filterByKey,
-    options: [...new Set(tracks?.map((data) => data[filterByKey]))],
-    clicked: false,
-  }
-
-  return filter
-}
